Refactor advert route imports to align with updated file naming conventions.
