+++ api/index.ts
@@ -1,6 +1,6 @@
 import express from 'express';
-import {userRouter} from './src/routes/userRoutes';
-import {advertRouter} from './src/routes/advertRoutes';
+import {userRouter} from './src/Routes/UserRoutes';
+import {advertRouter} from './src/Routes/RentalRoutes';
 import { serve, setup } from 'swagger-ui-express';
 import dotenv from 'dotenv';
 import { parseDocument } from 'yaml';
