+++ api/package.json
@@ -12,6 +12,7 @@
     "@prisma/adapter-pg": "^6.0.1",
     "@prisma/client": "^6.0.1",
     "bcrypt": "^5.1.1",
+    "cors": "^2.8.5",
     "express": "^4.21.1",
     "jsonwebtoken": "^9.0.2",
     "pg": "^8.13.1",
@@ -23,8 +24,8 @@
     "winston": "^3.17.0"
   },
   "devDependencies": {
-    "@types/cors": "^2.8.17",
     "@types/bcrypt": "^5.0.2",
+    "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
     "@types/jsonwebtoken": "^9.0.7",
     "@types/node": "^22.10.1",
