+++ api/src/Models/schema.prisma
@@ -5,7 +5,7 @@ datasource db {
 
 generator client {
   provider        = "prisma-client-js"
-  previewFeatures = ["relationJoins"]
+  previewFeatures = ["relationJoins", "omitApi"]
   output          = "./generated/prisma-client-js"
 }
