+++ client/tailwind.config.js
@@ -1,4 +1,5 @@
 /** @type {import('tailwindcss').Config} */
+
 export default {
   content: [],
   theme: {
@@ -6,4 +7,3 @@ export default {
   },
   plugins: [],
 }
-
